Rewrite logic

+++ Ajedrez/Clock.cpp
@@ -12,14 +12,16 @@ ChessClock::ChessClock(double initial_time, double increment)
 /// @brief Updates the col color clock
 /// @param col The color of the player wich clock you want to update
 /// @return A 0-1 float as % of remaining time from the initial time.Returns -1 if run out of time
-float	ChessClock::updateChessClock()
+float ChessClock::updateChessClock()
 {
-	if (m_total_time <= 0 || m_remaining > m_total_time) return 1;
-	if (m_remaining <= 0) return 0;
+	if (m_total_time <= 0 || m_remaining > m_total_time)
+		return 1;
+	if (m_remaining <= 0)
+		return 0;
 	m_remaining += m_increment;
 	m_remaining -= (float)(clock() - m_origin) / CLOCKS_PER_SEC;
 	m_origin = clock();
-	if (m_remaining <= 0) return 0;
-	return (m_remaining / m_total_time);
+	if (m_remaining <= 0)
+		return 0;
+	return (float)(m_remaining / m_total_time);
 }
-
